Enlarge potential size of modals for large screens
Change modals for large screens from 10% position from top and 80% max
height, to 5% position from top and 90% max height.

This is to accommodate the taller cookies policy modal.

It seems that it is not easy to apply a custom position and height to a
specific modal only. This doesn't seem to be supported by the modal
library currently being used.

Before the recent changes, most modals were rendering at 10px top
position (not the originally planned 10%), so changing this to 5%
doesn't seem to be much of a compromise.